feat: 네일샵 저장 및 취소 예외처리 추가

+++ src/main/java/backend/sumnail/domain/user/controller/dto/UserController.java
@@ -37,7 +37,7 @@ public class UserController {
 
     @PostMapping("nail-shops/{nailShopId}")
     public ResponseEntity<Void> saveNailShopUser(@PathVariable("nailShopId") long nailShopId) {
-        userService.saveNailShopUser((long) 1, nailShopId);
+        userService.saveNailShopUser(1, nailShopId);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
